Inline store.ts methods into persisted/index.ts
